+++ CHANGELOG.md
@@ -7,7 +7,7 @@ v1.5.13 (16-01-2016)
 - Merge PR (#170)[https://github.com/dirkgroenen/mopidy-mopify/pull/170]
 - Add stations to tracklist instead queue (#174)[https://github.com/dirkgroenen/mopidy-mopify/issues/174]
 - Automatically restore broken socket connection
-- 
+- Get queue length before inserting (#141)[https://github.com/dirkgroenen/mopidy-mopify/issues/141]
 
 v1.5.12 (27-12-2015)
 --------------------